Switched from Amazon Linux to Ubuntu 24.04 LTS for the AWS instance.
Working out which AMI code to use is a pig of a process.
I tried various different approaches unsuccessfully before finding this third-party module, which is very simple.
Maybe better to eliminate that dependency in the end, but it's simple and works.
Changed the root block device type from gp2 to gp3, as is now default in Ubuntu 24.04 LTS.